Cleaned up all legacy code, replaced classes with structs.
