+++ patches/tModLoader/Terraria/ModLoader/Default/LegacyUnloadedTilesSystem.cs
@@ -5,96 +5,42 @@ using Terraria.ModLoader.IO;
 
 namespace Terraria.ModLoader.Default
 {
-	[LegacyName("MysteryTilesWorld")]
-	class UnloadedTilesSystem : ModSystem
+	/// <summary>
+	/// Handles conversion of legacy 1.3 world unloaded tile TML TagCompound data to the newer 1.4+ systems.
+	/// </summary>
+	[LegacyName("MysteryTilesWorld", "UnloadedTilesSystem")]
+	internal partial class LegacyUnloadedTilesSystem : ModSystem
 	{
-		internal static List<LegacyUnloadedTileInfo> infos = new List<LegacyUnloadedTileInfo>();
-		//internal List<LegacyUnloadedTileInfo> pendingInfos = new List<LegacyUnloadedTileInfo>();
+		private static readonly List<TileInfo> infos = new List<TileInfo>();
+		private static readonly Dictionary<int, ushort> converted = new Dictionary<int, ushort>();
 
 		public override void OnWorldLoad() {
 			infos.Clear();
-			//pendingInfos.Clear();
 			converted.Clear();
 		}
 
-		/*
-		public override TagCompound SaveWorldData() {
-			return new TagCompound {
-				["list"] = infos.Select(info => info?.Save() ?? new TagCompound()).ToList()
-			};
-		}
-		*/
-
 		public override void LoadWorldData(TagCompound tag) {
-			//List<ushort> canRestore = new List<ushort>();
-			//bool canRestoreFlag = false;
-
 			foreach (var infoTag in tag.GetList<TagCompound>("list")) {
 				if (!infoTag.ContainsKey("mod")) {
-					infos.Add(null);
-					//canRestore.Add(0);
+					infos.Add(TileInfo.Invalid);
 					continue;
 				}
 
 				string modName = infoTag.GetString("mod");
 				string name = infoTag.GetString("name");
 				bool frameImportant = infoTag.ContainsKey("frameX");
-				var info = frameImportant ?
-					new LegacyUnloadedTileInfo(modName, name, infoTag.GetShort("frameX"), infoTag.GetShort("frameY")) :
-					new LegacyUnloadedTileInfo(modName, name);
+
+				var info = frameImportant
+					? new TileInfo(modName, name, infoTag.GetShort("frameX"), infoTag.GetShort("frameY"))
+					: new TileInfo(modName, name);
 
 				infos.Add(info);
-
-				//int type = ModContent.TryFind(modName, name, out ModTile tile) ? tile.Type : 0;
-
-				//canRestore.Add((ushort)type);
-
-				//if (type != 0)
-				//	canRestoreFlag = true;
 			}
 
 			if (infos.Count > 0) {
 				ConvertTiles();
 			}
-
-			/*if (canRestoreFlag) {
-				RestoreTiles(canRestore);
-
-				for (int k = 0; k < canRestore.Count; k++) {
-					if (canRestore[k] > 0) {
-						infos[k] = null;
-					}
-				}
-			}
-
-			if (pendingInfos.Count > 0) {
-				ConfirmPendingInfo();
-			}
-			*/
 		}
-
-		/*
-		private void RestoreTiles(List<ushort> canRestore) {
-			ushort unloadedType = ModContent.Find<ModTile>("ModLoader/UnloadedTile").Type;
-			for (int x = 0; x < Main.maxTilesX; x++) {
-				for (int y = 0; y < Main.maxTilesY; y++) {
-					if (Main.tile[x, y].type == unloadedType) {
-						Tile tile = Main.tile[x, y];
-						LegacyUnloadedTileFrame frame = new LegacyUnloadedTileFrame(tile.frameX, tile.frameY);
-						int frameID = frame.FrameID;
-						if (canRestore[frameID] > 0) {
-							LegacyUnloadedTileInfo info = infos[frameID];
-							tile.type = canRestore[frameID];
-							tile.frameX = info.frameX;
-							tile.frameY = info.frameY;
-						}
-					}
-				}
-			}
-		} 
-		*/
-
-		internal static Dictionary<int, ushort> converted = new Dictionary<int, ushort>();
 
 		internal void ConvertTiles() {
 			var legacyEntries = TileIO.Tiles.entries.ToList();
@@ -105,16 +51,19 @@ namespace Terraria.ModLoader.Default
 			for (int x = 0; x < Main.maxTilesX; x++) {
 				for (int y = 0; y < Main.maxTilesY; y++) {
 					Tile tile = Main.tile[x, y];
+
 					if (!(tile.active() && tile.type >= TileID.Count))
 						continue;
 
 					ushort type = tile.type;
+
 					if (TileIO.Tiles.entries[type] == null) { // Is an unloaded block
 						type = unloadedReader.Get(x, y);
 
 						if (legacyEntries[type].modName.Equals("ModLoader")) { // Was saved as an unloaded block
 							ConvertTile(tile, legacyEntries, out type);
 						}
+
 						builder.Add(x, y, type);
 					}
 
@@ -126,54 +75,22 @@ namespace Terraria.ModLoader.Default
 		}
 
 		internal void ConvertTile(Tile tile, List<TileEntry> entries, out ushort type) {
-			LegacyUnloadedTileFrame frame = new LegacyUnloadedTileFrame(tile.frameX, tile.frameY);
+			var frame = new TileFrame(tile.frameX, tile.frameY);
 			int frameID = frame.FrameID;
 
 			if (converted.TryGetValue(frameID, out type))
 				return;
 
-			LegacyUnloadedTileInfo info = infos[frameID];
-			var entry = new TileEntry(TileLoader.GetTile(tile.type));
-			entry.name = info.name;
-			entry.modName = info.modName;
-			entry.frameImportant = info.frameX > -1;
-			entry.type = type = (ushort)entries.Count;
+			TileInfo info = infos[frameID];
+			var entry = new TileEntry(TileLoader.GetTile(tile.type)) {
+				name = info.name,
+				modName = info.modName,
+				frameImportant = info.frameX > -1,
+				type = type = (ushort)entries.Count
+			};
 
 			entries.Add(entry);
 			converted.Add(frameID, type);
 		}
-
-		/*
-		private void ConfirmPendingInfo() {
-			List<int> truePendingID = new List<int>();
-			int nextID = 0;
-			for (int k = 0; k < pendingInfos.Count; k++) {
-				while (nextID < infos.Count && infos[nextID] != null) {
-					nextID++;
-				}
-				if (nextID == infos.Count) {
-					infos.Add(pendingInfos[k]);
-				}
-				else {
-					infos[nextID] = pendingInfos[k];
-				}
-				truePendingID.Add(nextID);
-			}
-			ushort pendingType = ModContent.Find<ModTile>("ModLoader/PendingUnloadedTile").Type;
-			ushort unloadedType = ModContent.Find<ModTile>("ModLoader/UnloadedTile").Type;
-			for (int x = 0; x < Main.maxTilesX; x++) {
-				for (int y = 0; y < Main.maxTilesY; y++) {
-					if (Main.tile[x, y].type == pendingType) {
-						Tile tile = Main.tile[x, y];
-						LegacyUnloadedTileFrame frame = new LegacyUnloadedTileFrame(tile.frameX, tile.frameY);
-						frame = new LegacyUnloadedTileFrame(truePendingID[frame.FrameID]);
-						tile.type = unloadedType;
-						tile.frameX = frame.FrameX;
-						tile.frameY = frame.FrameY;
-					}
-				}
-			}
-		}
-		*/
 	}
 }
